Add files via upload
The Python file Logistic_Regression offers an insight to the basic statistics involved while calculating logistic regression for the dataset ex2data1.txt. To begin with an appropriate cost function is made with the help of error functions(difference between the sigmoid for a given value of theta and the actual value of y).Then this cost function is repeatedly minimized with the help of gradient descent technique.Finally the value of theta selected is used to calculate the probability of being positive for each case.By assigning a certain threshold to that probability, 1 or 0 is set at y for a particular record.Finally, the accuracy of the predictions is calculated by comparing with the actual values of y given in the data set.
This solution however suffers from the problem of overfitting which we intend to reesolve in our next upload.